chore: same thing for `rpcfuzz`
