+++ doc/polycli_rpcfuzz.md
@@ -14,7 +14,7 @@
 Continually run a variety of RPC calls and fuzzers.
 
 ```bash
-polycli rpcfuzz http://localhost:8545 [flags]
+polycli rpcfuzz [flags]
 ```
 
 ## Usage
@@ -114,6 +114,7 @@ $  docker run -v $PWD/contracts:/contracts ethereum/solc:stable --storage-layout
       --md                        Flag to indicate that output will be exported as a Markdown.
       --namespaces string         Comma separated list of rpc namespaces to test (default "eth,web3,net,debug,raw")
       --private-key string        The hex encoded private key that we'll use to sending transactions (default "42b6e34dc21598a807dc19d7784c71b2a7a01f6480dc6f58258f78e539f1a1fa")
+  -r, --rpc-url string            The RPC endpoint url (default "http://localhost:8545")
       --seed int                  A seed for generating random values within the fuzzer (default 123456)
 ```
